Parity,AND XOR operator
In this problem, if the answer is not Alice, it must be Bob (Jury guarantees). Vice versa.

The + and ⊕ operations have one thing in common: the parity of the res of A + B or A ⊕ B will be opposite with A only if B is an odd number. Visually:

A	B	A+B and A ⊕ B are both
odd	odd	even
odd	even	odd
even	odd	odd
even	even	even
So we can count the odd numbers in the input array. If the number of odd numbers is an odd number, the parity of the final result will be opposite to the starting.

If the number of odd numbers is odd, then the parity of the starting number and ending number must be different. So if x and y have the same parity, it's impossible to let Alice win. Then the answer will be Bob.

On the contrary, if the number of odd numbers is even, the parity of the starting number and ending number must be the same. So if x and y have different parity, it's also impossible to let Alice win. Therefore, the answer is also Bob.

For other conditions, Alice will win.